JUS-25: refactor: Cleanup, fix constants and exports, remove logs
